refactor: update pagination comments and change WhereNot function to use Not method
