--- gorm/scope/compose.go
@@ -1,5 +1,0 @@
-package scope
-
-import "gorm.io/gorm"
-
-type Scopes []func(*gorm.DB) *gorm.DB
