--- gorm/scope/trait.go
@@ -1,1 +1,0 @@
-package scope
