--- gorm/scope/where.go
@@ -1,7 +1,0 @@
-package scope
-
-import (
-	"fmt"
-
-	"gorm.io/gorm"
-)
